python-toolkit/pyproject.toml
[project]
name = "python-toolkit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pytest>=8.4.1",
    "python-dateutil>=2.9.0.post0",
    "python-docx>=1.2.0",
]
